backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
oracledb
python-dotenv
gunicorn

oracledb
firebase_admin